requirements.txt
# --- Framework & Server ---
fastapi==0.135.1
uvicorn==0.42.0
python-multipart==0.0.22
pydantic-settings==2.2.1

# --- AI & NLP ---
spacy==3.7.4  
openai==2.29.0 

# --- PDF Processing ---
pymupdf==1.23.26

# --- Infrastructure & Utilities ---
httpx==0.28.1
requests==2.31.0
python-dotenv==1.0.1